Create chatBot.java
This is a chat bot project that i found on JetBrains Academy. The bot can accomplish simple tasks such as introducing itself, asking for user's name, guessing user's age based on the remainders of three operations and it can ask a single question until the user gets it right (while (!=)).